# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Gestione Dispensa Public API
  version: 2.0.0
  description: Canonical browser-facing API exposed by the Gateway. Domain services remain internal network boundaries.
servers:
  - url: https://<host>:8443/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post: {tags: [Identity], operationId: register, requestBody: {$ref: '#/components/requestBodies/Json'}, responses: {'201': {$ref: '#/components/responses/Data'}}}
...
  /inventory/stock-items:
    get: {tags: [Inventory], operationId: listStockItems, responses: {'200': {$ref: '#/components/responses/Data'}}}
    post: {tags: [Inventory], operationId: createStockItem, parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/Json'}, responses: {'201': {$ref: '#/components/responses/Data'}}}
  /inventory/stock-items/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Inventory], operationId: getStockItem, responses: {'200': {$ref: '#/components/responses/Data'}}}
    patch: {tags: [Inventory], operationId: updateStockItem, parameters: [{$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}], requestBody: {$ref: '#/components/requestBodies/Json'}, responses: {'200': {$ref: '#/components/responses/Data'}}, '409': {$ref: '#/components/responses/Error'}}
    delete: {tags: [Inventory], operationId: deleteStockItem, parameters: [{$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'204': {description: Deleted}}}
...
  /products/{id}:
    parameters: [{$ref: '#/components/parameters/Id'}]
    get: {tags: [Catalog], operationId: getProduct, responses: {'200': {$ref: '#/components/responses/Data'}}}
    patch: {tags: [Catalog], operationId: updateProduct, requestBody: {$ref: '#/components/requestBodies/Json'}, responses: {'200': {$ref: '#/components/responses/Data'}}}
...
  /notifications/{id}/read:
    parameters: [{$ref: '#/components/parameters/Id'}]
    post: {tags: [Notifications], operationId: markNotificationRead, parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {$ref: '#/components/responses/Data'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
    IfMatch: {name: If-Match, in: header, required: false, schema: {type: string}}
    IdempotencyKey: {name: X-Idempotency-Key, in: header, required: false, schema: {type: string, minLength: 8, maxLength: 256}}
  requestBodies:
    Json:
      required: true
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
  responses:
    Data:
      description: Successful response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/DataEnvelope'}
...
    Error:
      description: Error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
  schemas:
    DataEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta: {type: object, additionalProperties: true}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: {type: string}
            message: {type: string}
            retryable: {type: boolean}
        meta: {type: object, additionalProperties: true}