# services/service-identity/src/server.ts
import express from "express"; import { PostgresClient,resolveDatabaseUrl } from "./db/postgres-client.js"; import { OidcTokenVerifier } from "./identity/oidc.js"; import { PostgresUserProfileRepository } from "./identity/users.js"; import { parseRegisterUserInput,registerUser,resolveKeycloakAdminConfig,requestPasswordReset } from "./identity/register.js"; import { buildMeta,sendFailure,sendSuccess } from "./http/envelope.js"; import { corsMiddleware,requestMetaMiddleware,resolvePrincipal } from "./http/middleware.js";
const port=Number(process.env.PORT??3310),app=express();app.use(corsMiddleware());app.use(requestMetaMiddleware());app.use(express.json({limit:'2mb'}));const pg=PostgresClient.create({connectionString:resolveDatabaseUrl()});const verifier=await OidcTokenVerifier.fromIssuer(process.env.OIDC_ISSUER!,process.env.OIDC_AUDIENCE!,fetch,{...(process.env.OIDC_DISCOVERY_URL ? {discoveryUrl:process.env.OIDC_DISCOVERY_URL} : {}),...(process.env.OIDC_JWKS_URL ? {jwksUrl:process.env.OIDC_JWKS_URL} : {})});const profiles=new PostgresUserProfileRepository(pg);
app.get('/health/live',(_q,res)=>res.json({status:'ok',service:'service-identity'}));app.get('/health/ready',async(_q,res)=>{try{await pg.ping();res.json({status:'ready'})}catch{res.status(503).json({status:'not_ready'})}});
app.get('/api/v1/meta',(req,res)=>sendSuccess(res,200,{name:'service-identity',version:process.env.APP_VERSION??'0.1.0',contract:'api/v1'},req.meta));
app.get('/api/v1/me',async(req,res)=>{const p=await resolvePrincipal(req,verifier);if(!p)return sendFailure(res,401,'UNAUTHENTICATED','Authentication is required.',req.meta);try{await profiles.upsertFromOidc({id:p.subject,...(p.email?{email:p.email}:{}),...(p.name?{displayName:p.name}: {})});}catch{} const familyId=await profiles.getActiveFamilyId(p.subject).catch(()=>undefined);sendSuccess(res,200,{id:p.subject,subject:p.subject,issuer:p.issuer,roles:p.roles,scopes:p.scopes,expiresAt:p.expiresAt.toISOString(),...(p.email?{email:p.email}:{}),...(p.name?{name:p.name}:{}),...(familyId?{activeFamilyId:familyId}: {})},req.meta);});
app.get('/api/v1/auth/me',async(req,res)=>{const p=await resolvePrincipal(req,verifier);if(!p)return sendFailure(res,401,'UNAUTHENTICATED','Authentication is required.',req.meta);sendSuccess(res,200,{id:p.subject,subject:p.subject,email:p.email,name:p.name,roles:p.roles,scopes:p.scopes},req.meta)});
app.post('/api/v1/auth/register',async(req,res)=>{const parsed=parseRegisterUserInput(req.body);if(!parsed)return sendFailure(res,400,'VALIDATION_ERROR','Nome, email valida e password di almeno 8 caratteri sono obbligatori.',req.meta);try{sendSuccess(res,201,await registerUser(parsed,resolveKeycloakAdminConfig()),req.meta)}catch(e){sendFailure(res,400,'REGISTRATION_FAILED',e instanceof Error?e.message:'Registration failed',req.meta)}});
app.post('/api/v1/auth/reset-password',async(req,res)=>{const email=typeof req.body?.email==='string'?req.body.email.trim().toLowerCase():'';if(!email)return sendFailure(res,400,'VALIDATION_ERROR','A valid email is required.',req.meta);await requestPasswordReset(email,resolveKeycloakAdminConfig());sendSuccess(res,202,{accepted:true},req.meta)}); app.post('/api/v1/auth/logout',(_q,res)=>res.status(204).end());
app.use((req,res)=>sendFailure(res,404,'NOT_FOUND_OR_NOT_VISIBLE','The resource is not available.',req.meta??buildMeta(req)));app.listen(port,'0.0.0.0',()=>console.log(JSON.stringify({service:'service-identity',port})));